requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-admin==1.0.4
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_resources==6.5.2
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pendulum==3.1.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
setuptools==80.3.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn